move common code to base class
